+++ server/models/response/index.ts
@@ -1,0 +1,7 @@
+export interface APIPageResponse<T> {
+    count: number,
+    next:string | null | undefined,
+    previous: string | null | undefined, 
+    results: T[]
+}
+
