+++ server/package.json
@@ -12,9 +12,11 @@
   "dependencies": {
     "axios": "^1.7.3",
     "axios-cache-interceptor": "^1.5.3",
+    "express": "^4.19.2",
     "pokenode-ts": "^1.20.0"
   },
   "devDependencies": {
+    "@types/express": "^4.17.21",
     "@types/http-server": "^0.12.4",
     "@types/node": "^22.0.0",
     "nodemon": "^3.1.4",
